refactor: move specific template tests to their respective modules
