+++ packages/template/typescript/package.json
@@ -17,5 +17,8 @@
     "@electron-forge/async-ora": "6.0.0-beta.46",
     "@electron-forge/shared-types": "6.0.0-beta.46",
     "fs-extra": "^8.1.0"
+  },
+  "devDependencies": {
+    "@electron-forge/core": "6.0.0-beta.46"
   }
 }
